% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{add_metalab_summary_info}
\alias{add_metalab_summary_info}
\title{Add metalab summary info used by Metalab}
\usage{
add_metalab_summary_info(metalab_dataset_info, metalab_data)
}
\arguments{
\item{metalab_dataset_info}{, metalab_data}
}
\value{
Null
}
\description{
Add metalab summary info used by Metalab
}
