% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{get_metalab_dataset_info}
\alias{get_metalab_dataset_info}
\title{Get raw data used by Metalab}
\usage{
get_metalab_dataset_info(
  dataset_file = paste0(ml_metadata_url, "datasets.yaml")
)
}
\arguments{
\item{dataset_file}{A file or URL with MetaLab dataset metadata, defaults to the current MetaLab dataset metadata hosted in the langcog/metalab2 repository on Github}
}
\value{
A data.frame of MetaLab dataset metadata
}
\description{
Get raw data used by Metalab
}
\examples{
\dontrun{
  ml_dataset_info <- metalabr::get_metalab_dataset_info()
  ml_data <- metalabr::get_metalab_data(ml_dataset_info)
}

}
