% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_and_validate_dataset}
\alias{load_and_validate_dataset}
\title{Load and validate dataset used by Metalab}
\usage{
load_and_validate_dataset(dataset_info, field_info)
}
\arguments{
\item{dataset_info}{A data.frame of datasets created with get_metalab_dataset_info()}
}
\value{
Boolean whether dataset is valid
}
\description{
Load and validate dataset used by Metalab
}
