% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalabr-package.R
\docType{package}
\name{metalabr-package}
\alias{metalabr}
\alias{metalabr-package}
\title{metalabr: access MetaLab community-augmented meta-analysis data}
\description{
Read released, versioned MetaLab data (effect sizes from meta-analyses of
language acquisition and cognitive development), validate candidate
datasets against the MetaLab field specification, and draw the standard
MetaLab visualizations. See \url{https://metalab.stanford.edu}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/langcog/metalabr}
  \item \url{https://metalab.stanford.edu}
  \item Report bugs at \url{https://github.com/langcog/metalabr/issues}
}

}
\author{
\strong{Maintainer}: Michael C. Frank \email{mcfrank@stanford.edu} (\href{https://orcid.org/0000-0002-7551-4378}{ORCID})

Authors:
\itemize{
  \item Erik Iverson \email{erik@sigmafield.org}
  \item Sara El-Shawa \email{sara.el.shawa@utoronto.ca}
}

Other contributors:
\itemize{
  \item MetaLab contributors [copyright holder]
}

}
\keyword{internal}
