% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metalab_data.R
\name{get_metalab_data}
\alias{get_metalab_data}
\title{Get raw data used by Metalab}
\usage{
get_metalab_data(ml_metadata, short_names, domains, specs)
}
\arguments{
\item{dataset_info}{A data.frame of datasets created with get_metalab_dataset_info()}
}
\value{
A data.frame of raw data read from Google Sheets
}
\description{
Get raw data used by Metalab
}
\examples{
\dontrun{
  ml_metadata <- metalabr::get_ml_metadata()
  ml_data <- metalabr::get_ml_data(ml_metadata)
}

}
