% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{get_metalab_data}
\alias{get_metalab_data}
\title{Get raw data used by Metalab}
\usage{
get_metalab_data(dataset_info)
}
\arguments{
\item{dataset_info}{A data.frame of datasets created with get_metalab_dataset_info()}
}
\value{
A data.frame of raw data read from Google Sheets
}
\description{
Get raw data used by Metalab
}
\examples{
\dontrun{
  ml_dataset_info <- metalabr::get_metalab_dataset_info()
  ml_data <- metalabr::get_metalab_data(ml_dataset_info)
}

}
