% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metalab_metadata.R
\name{get_metalab_metadata}
\alias{get_metalab_metadata}
\title{Get raw data used by Metalab}
\usage{
get_metalab_metadata(
  dataset_file = paste0(metalab_metadata_url, "datasets.yaml")
)
}
\arguments{
\item{dataset_file}{A file or URL with MetaLab dataset metadata, defaults to the current MetaLab dataset metadata hosted in the langcog/metalab2 repository on Github}
}
\value{
A data.frame of MetaLab dataset metadata
}
\description{
Get raw data used by Metalab
}
\examples{
\dontrun{
  metalab_dataset_info <- metalabr::get_metalab_dataset_info()
  metalab_data <- metalabr::get_metalab_data(metalab_dataset_info)
}

}
